requirements.txt
picamera==1.13
alipay_sdk_python==3.3.398
pyzbar==0.1.8
